Rename dialog module to ajust to YaST convention
