+++ keyboard_rewrite/src/clients/keyboard.rb
@@ -1,3 +1,3 @@
-require_relative "../lib/y2_keyboard/dialog/layout_selector"
+require_relative "../lib/y2_keyboard/dialogs/layout_selector"
 
-Y2Keyboard::Dialog::LayoutSelector.run
+Y2Keyboard::Dialogs::LayoutSelector.run
